refactor: integrate two react-resval version
