+++ packages/react-resval/src/index.ts
@@ -1,2 +1,9 @@
+// Notes: Build for React Resval v1.0.0
+// Export for test only in development mode and comment out the following line if you don't want to include in the build process.
 export * as v1 from './v1.0.0'
+// Export for build in production mode.
+// export * from './v1.0.0'
+
+// Notes: Build for React Resval v0.1.0
+// Comment out the following line if you don't want to include in the build process.
 export * from './v0.1.0'
